refactor: streamline hook imports in MarketDetailV2
- Updated MarketDetailV2 to import useAutoRefreshTokenDetail from a centralized hooks index file, enhancing code organization and readability.
- Added useAutoRefreshTokenDetail export to the hooks index file for better accessibility across components.
