+++ packages/kit/src/views/Market/MarketDetailV2/hooks/index.ts
@@ -1,4 +1,4 @@
-export * from './useMarketDetail';
 export * from './useTokenDetail';
 export * from './useMarketHolders';
 export * from './useMarketTransactions';
+export * from './useAutoRefreshTokenDetail';
